+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 2.0.0
+- Update native SDK dependencies: 4.16.0 for iOS and 4.16.+ for Android.
+- Add new method setEmail to set user email.
+- Add new method setUserProfile to set customerId, email and tags in a single call.
+- Add new method trackUserTags to track user tags.
+- Add new method trackPurchase to track purchases with products in a single call.
+- Add new method unregisterPushSystem to unregister from push notifications.
+- Deprecate trackExtraUserInfo method.
+- Deprecate startOrder, addProduct and trackOrder methods.
+- Remove cancelOrder method.
+
 ## 1.9.2
 - Update native SDK dependencies: 4.15.7 for Android and 4.15.6 for iOS.